+++ tests/checks/test_ast_imports.py
@@ -49,6 +49,14 @@ FTP015 = partial(
         "which should be replaced with a proper alternative of importlib.*"
     ),
 )
+_FTP133 = partial(
+    Issue,
+    issue_number="FTP133",
+    message=(
+        "Using the compression namespace is recommended. "
+        "Replace the imported module with compression.{module}"
+    ),
+)
 
 
 def FTP001(  # pylint:disable=invalid-name
@@ -84,6 +92,13 @@ def FTP030(  # pylint:disable=invalid-name
 ) -> Issue:
     issue = _FTP030(line=line, column=column)
     return issue._replace(message=issue.message.format(future=future))
+
+
+def FTP133(  # pylint:disable=invalid-name
+    *, line: int, column: int, module: str
+) -> Issue:
+    issue = _FTP133(line=line, column=column)
+    return issue._replace(message=issue.message.format(module=module))
 
 
 def test_ftp001(runner: Flake8RunnerFixture) -> None:
@@ -190,3 +205,39 @@ def test_ftp058(runner: Flake8RunnerFixture) -> None:
         FTP058(line=10, column=1),
         FTP058(line=11, column=1),
     ]
+
+
+@pytest.mark.parametrize(
+    "python_version,find_by_version", [("3.7.1", False), ("3.14.1", True)]
+)
+@pytest.mark.parametrize(
+    "imp,find_by_imp,module",
+    [
+        ("import bz2", True, "bz2"),
+        ("import gzip", True, "gzip"),
+        ("import lzma", True, "lzma"),
+        ("import zlib", True, "zlib"),
+        ("from zlib import some", True, "zlib"),
+        ("import zlib2", False, None),
+        ("from compression.gzip import some", False, None),
+        ("import compression.lzma", False, None),
+    ],
+)
+def test_ftp133(
+    runner: Flake8RunnerFixture,
+    python_version: str,
+    find_by_version: bool,
+    imp: str,
+    find_by_imp: bool,
+    module: str,
+) -> None:
+    results = runner(
+        filename="ftp133.txt",
+        issue_number="FTP133",
+        imp=imp,
+        args=("--ftp-python-version", python_version),
+    )
+    if find_by_version and find_by_imp:
+        assert results == [FTP133(line=1, column=1, module=module)]
+    else:
+        assert not results